Abstract white/blacklist parsing logic
